+++ src/components/rows/EdgeRow.tsx
@@ -15,6 +15,7 @@ import { triggerHaptic } from '../../util/haptic'
 import { fixSides, mapSides, sidesToMargin } from '../../util/sides'
 import { EdgeTouchableOpacity } from '../common/EdgeTouchableOpacity'
 import { ChevronRightIcon } from '../icons/ThemedIcons'
+import { ShimmerText } from '../progress-indicators/ShimmerText'
 import { showToast } from '../services/AirshipInstance'
 import { cacheStyles, type Theme, useTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
@@ -39,6 +40,7 @@ interface Props {
   error?: boolean
   icon?: React.ReactNode
   loading?: boolean
+  shimmer?: boolean
   maximumHeight?: 'small' | 'medium' | 'large'
   rightButtonType?: RowActionIcon
   title?: string
@@ -57,6 +59,7 @@ export const EdgeRow: React.FC<Props> = (props: Props) => {
     error,
     icon,
     loading = false,
+    shimmer = false,
     marginRem,
     maximumHeight = 'medium',
     testID,
@@ -123,7 +126,9 @@ export const EdgeRow: React.FC<Props> = (props: Props) => {
             {title}
           </EdgeText>
         )}
-        {loading ? (
+        {shimmer ? (
+          <ShimmerText characters={title?.length} />
+        ) : loading ? (
           <ActivityIndicator
             style={styles.loader}
             color={theme.primaryText}